incorporate requested changes
